+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Fix bug when updating security callbacks
+* Fix doubling of `size_is` for `NdrConformantString`
 
 
 ## v1.2.1 - Mar 29, 2025