+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* Refactor checks for known structs
 * Fix bug when updating security callbacks
 * Fix doubling of `size_is` for `NdrConformantString`