+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## v1.2.2 - Apr XX, 2025
 
+### Added
+
+* Add support for pointer conformance
+
 ### Changed
 
 * Refactor checks for known structs