+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix bug when updating security callbacks
 * Fix doubling of `size_is` for `NdrConformantString`
 * Fix incorrect offset in NDR variable expressions
+* Fix incorrect handling of teneray operator in NDR expressions
 
 
 ## v1.2.1 - Mar 29, 2025